Rename main.py to news.py and update all imports
- Renamed main.py to news.py for better module naming
- Updated all import statements across 7 files:
  - README.md
  - examples.py
  - quick_streaming_test.py
  - simple_streaming_test.py
  - simple_test.py
  - test_batch_news.py
  - test_streaming_news.py
- All imports changed from 'from main import NewsDataSource' to 'from news import NewsDataSource'
- Verified functionality with both batch and streaming tests
- Tests passed successfully confirming rename operation completed without issues
