+++ src/py/data_source/news/simple_streaming_test.py
@@ -10,7 +10,7 @@ import os
 import time
 from pyspark.sql import SparkSession
 from pyspark.sql.functions import col, current_timestamp
-from main import NewsDataSource
+from news import NewsDataSource
 
 
 def test_simple_streaming():
